Extract chart card recommendations
