+++ tests/test-recommendations.js
@@ -40,6 +40,7 @@ globalThis.fetch = async (url, opts) => {
   const mainSrc = await fetchWithRetry('js/main.js');
   const chatSrc = await fetchWithRetry('js/chat.js');
   const viewsSrc = await fetchWithRetry('js/views.js');
+  const chartCardRecsSrc = await fetchWithRetry('js/chart-card-recs.js');
   const markerDetailSrc = await fetchWithRetry('js/marker-detail-modal.js');
   const dashboardWidgetsSrc = await fetchWithRetry('js/dashboard-widgets.js');
   const contextSrc = await fetchWithRetry('js/context-cards.js');
@@ -199,7 +200,8 @@ globalThis.fetch = async (url, opts) => {
   assert('chat.js has rec-chat-wrapper class', chatSrc.includes('rec-chat-wrapper'));
   assert('views.js has chart-rec placeholder in header', viewsSrc.includes('chart-rec-'));
   assert('views.js keeps chart title text separate from tips host', viewsSrc.includes('chart-card-title-text') && viewsSrc.includes('chart-card-tips-host'));
-  assert('views.js has loadChartCardRecs function', viewsSrc.includes('function loadChartCardRecs'));
+  assert('views.js imports chart card recommendation module', viewsSrc.includes("from './chart-card-recs.js'"));
+  assert('chart-card-recs.js has loadChartCardRecs function', chartCardRecsSrc.includes('function loadChartCardRecs'));
   assert('marker-detail-modal.js scrollToRec auto-opens details', markerDetailSrc.includes('scrollToRec'));
   assert('loadCatalog on window', typeof window.loadCatalog === 'function');
   assert('nav.js exposes recommendations sidebar helper', navSrc.includes('openRecommendationsFromSidebar'));
